requirements.txt
-f https://download.pytorch.org/whl/cpu/torch_stable.html
opencv-python-headless
dash
gunicorn
requests
plotly
torch==1.10.0+cpu
torchvision==0.11.1+cpu
numpy
Pillow
